Catalog or calculate package metrics

When no object is passed, returns a list of possible metrics. When a pkg object is provided, return metrics calculated for that package. Metrics are a subset of all the data calculated over the course of assessing a package. For access to all the internally calculated data, pass all = TRUE.

Usage

metrics(x, ..., all = FALSE)

Arguments

x

Optionally, an object to retrieve metrics from. When NULL (the default), a listing of metric metadata is returned.

...

Additional arguments unused.

all

If TRUE, include non-metric package data. These are often intermediate data used in the calculation of package metrics.

Value

A list of calculated values or metadata, in the cases where an object is or is not provided respectively.

Metrics

The following metrics are provided by val.meter.

R CMD check Error Count

<integer> the number of errors produced when running R CMD check

Total Downloads

<integer> total number of lifetime downloads, as reported by the Posit CRAN mirror through the cranlogs API

Dependency Count

<integer> the number of required dependencies

Has Website

<logical> a logical indicating whether the package has a website