Create a new package object to start measuring package heuristics

Aggregates package information, surveying metadata from discovered resources.

Usage

pkg(resource, permissions, policy = opt("policy"))

Arguments

resource

resource (often a multi_resource), providing the resources to be used for deriving packages data. If a multi_resource, the order of resources determines the precedence of information. If information about a package could be derived from multiple sources, the first source is prioritized.

permissions

permissions granted for deriving data. If not provided, the default from policy will be used.

policy

policy to use when converting input to resources. Most commonly used for interpreting strings as resources.